added global stats

+++ researchers.py
@@ -81,10 +81,10 @@ def get_researchers_info_from_project(project_id):
     global_stats['total_publications']=int(len(set(total_publications)))
     # add other publication stats
     df1=pd.read_csv("total_publications.csv")
-    global_stats['average_publications']=df1['count'].mean()
-    global_stats['max_publications']=df1['count'].max()
-    global_stats['min_publications']=df1['count'].min()
-    global_stats['std_dev_publications']=df1['count'].std()
+    global_stats['average_publications']=int(df1['count'].mean())
+    global_stats['max_publications']=int(df1['count'].max())
+    global_stats['min_publications']=int(df1['count'].min())
+    global_stats['std_dev_publications']=float(df1['count'].std())
 
     return data,global_stats
 
@@ -98,8 +98,8 @@ def get_project_info_investors(project_acronym=''):
     sub_df['ecContribution']=pd.to_numeric(sub_df['ecContribution'])
     # add avg sum investment
     df1=pd.read_csv("total_cost_projects.csv")
-    data['avg_ecContribution']=df1['sum'].mean()
-    data['std_dev_ecContribution']=df1['sum'].std()
+    data['avg_ecContribution']=float(df1['sum'].mean())
+    data['std_dev_ecContribution']=float(df1['sum'].std())
     
     vals=sub_df['ecContribution'].values.tolist()
     vals= [x for x in vals if ~np.isnan(x)] # remove nans
